teams/team_7/backend/requirements.txt
boto3>=1.35.0
pygame>=2.6.0
openai-whisper>=20231117
sounddevice>=0.4.6
numpy>=1.24.0
pydantic-settings>=2.0.0
chromadb>=1.0.0
langchain-aws>=0.2.0
langchain-chroma>=0.2.0
